Fixed up row table printing by giving column names
Previously, a row was printed out in a single column, i.e.
by converting the Python list to a string.

This change ensures that each column is printed individually.
For now this means auto-generating column names.  Once there's
an API call for grabbing a datasource schema, we will change
the datasource column names to meaningful strings.

Change-Id: I1c0f0c429a75284250ec28211504620e9eec05ad