Favor the use of importlib over Python internal __import__ statement
From the Python documentation, we have:

Because this function is meant for use by the Python interpreter and
not for general use it is better to use importlib.import_module() to
programmatically import a module.

This change implements exactly what the Python documentation, and
recognized best practices, recommend.

Also, move import_class utility function before its used in
get_client_class. Even though there are no behaviour differences
because of the order of definition of functions within the same Python
module, it helps users when reading the code.

Change-Id: I85e769e3ac108575a97351780bf0f5b3ef59cfb2
Signed-off-by: Cleber Rosa <cleber.gnu@gmail.com>