Revert "Fix the problem of default width is small"
In further testing, the default output without limit is unreadable on some consoles. The --fit-width option can be specified to fit to terminal width.

I think this is a new behavior that has to do with an openstackclient change that defaults fit-width to off. Reverting this change given the new situation.

This reverts commit 46e4bc630aa510c1689652f6a045067982021f51.

Change-Id: Ia43dd7e4dad8e8ca37d8822ffedaad750c73ba6c